requirements.txt
asn1crypto==0.24.0
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
cloud-init==18.5
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
distro-info===0.18ubuntu0.18.04.1
Django==1.11.11
djangorestframework==3.11.0
docker-py==1.10.6
docker-pycreds==0.4.0
Flask==1.1.1
gunicorn==20.0.4
httplib2==0.9.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lxml==4.4.2
MarkupSafe==1.1.1
netifaces==0.10.4
oauthlib==2.0.6
PAM==0.4.2
parsel==1.5.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pycrypto==2.6.1
PyDispatcher==2.0.5
pygobject==3.26.1
PyHamcrest==1.9.0
PyJWT==1.5.3
pyOpenSSL==19.1.0
pyserial==3.4
python-apt==1.6.4
python-debian==0.1.32
python-scrapyd-api==2.1.2
pytz==2018.3
pyxdg==0.25
PyYAML==5.2
queuelib==1.5.0
requests==2.22.0
requests-unixsocket==0.1.5
retrying==1.3.3
scrapinghub==2.3.0
Scrapy==1.8.0
scrapyd==1.2.1
SecretStorage==2.3.1
service-identity==18.1.0
shub==2.10.0
six==1.13.0
sqlparse==0.2.4
ssh-import-id==5.7
systemd-python==234
toml==0.10.0
tqdm==4.41.0
Twisted==19.10.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.7
virtualenv==16.7.9
w3lib==1.21.0
websocket-client==0.57.0
Werkzeug==0.16.0
zope.interface==4.7.1
